Refactor to abstract common code to common.ts
